pyproject.toml
[tool.poetry]
name = "jobtractor"
version = "0.1.0"
description = ""
authors = ["jstet <mail@jstet.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.324"
playwright = "^1.39.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
litellm = "^0.12.5"
pydantic = "^2.4.2"
rich = "^13.6.0"
dagster = "^1.5.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "jobtractor"